requirements.txt
Django==4.2
Pillow==9.5.0
django-model-utils==4.3.1
django-autoslug==1.9.9